Image hacks, finally over?
iOS image->canvas bug number 3:
If you give a width/height that go outside of the image boundary,
up to a certain point it appears to "clamp" it at the image dimensions.
Then apparently it just fails to draw completely.
That hid the fact that both boundaries actually need to be halved.